move: 테스트코드위치도 패키지와 똑같이 변경
