+++ src/test/java/bowling/domain/pins/PinsTest.java
@@ -1,5 +1,6 @@
-package bowling.domain;
+package bowling.domain.pins;
 
+import bowling.domain.Status;
 import bowling.domain.pins.Pins;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
